requirements.txt
Django==1.5.1
South==0.8.1
django-social-auth==0.7.24
oauth2==1.5.211
python-openid==2.2.5
simplejson==3.3.0
Unipath==1.0
coverage==3.6
django-discover-runner==1.0
selenium==2.33.0
mock==1.0.1
django-tracking2==0.1.4
django-widget-tweaks==1.3
